Remove use of GITHUB_TOKEN
According to https://github.com/actions/checkout#usage a token is already passed (and saved during the workflow run) through the checkout, so our workflow looks like it should already have that access.

The thing that I was thinking of was if we needed to pass the token to a custom action, but it looks like the action we're using to commit is just using bash, so perhaps it picks up that token in the git config? (I suppose we'll see).  Their [action.yml](https://github.com/stefanzweifel/git-auto-commit-action/blob/master/action.yml) at least does not indicate that they take a token as an input anyway.