Update tasks for Aardvark harvesting
This updates the main rake tasks to take account of our new
metadata format so that it can be harvested like we did before.

Also closes #143 by switching to purl-fetcher's API for asking
about what was released, which means we don't need to crawl the
whole site to determine deletions.